Fixes a bug in phosphpor
Phosphor gets confused when the type of a parameter slot changes.
This can happen when frames are generated by ASM instead of javac.
This is important when a primitive array is stored in the type
Object: to fix this we add another quick pass to always force
paramter variables of type java/lang/Object to have that type,
even when ASM has inferred a more specific type.